add image annotation in chart metadata (#46)
This pull request adds an annotation to the `charts/helm/Chart.yaml`
file to document the container image used for the
`licensing-example-java` chart. This helps clarify which image is
deployed and makes it easier for users and automated tooling to identify
the image reference.

* Chart metadata improvement:
* Added an `omnistrate.com/images` annotation specifying the container
image and tag for the `licensing-example-java` chart in `Chart.yaml`.